requirements.txt
anthropic
requests
python-dotenv
pathlib